+++ update.sh
@@ -2,7 +2,7 @@
 
 # Xavier Framework Update Script
 # Updates existing Xavier installations to the latest version
-# Version 1.1.10
+# Version 1.1.11
 
 set -e
 set -o pipefail